ASSIGNMENT CREATOR/requirements.txt
streamlit==1.33.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.4

# FIXED Gemini deps
google-generativeai==0.3.2
google-api-core==2.17.1
proto-plus==1.23.0
protobuf==4.25.3
typing-extensions>=4.9.0